Niveis de permissão e listagem de arquivos

+++ app/Http/Controllers/CampoController.php
@@ -12,7 +12,8 @@ use Illuminate\Support\Facades\Validator;
 class CampoController extends Controller
 {
 
-    public function salvarCampo(Request $request) {
+    public function salvarCampo(Request $request)
+    {
 
         $validator = Validator::make($request->all(), [
             "secao_id" => "required|exists:secaos,id",
@@ -21,61 +22,67 @@ class CampoController extends Controller
             "legenda" => "required|min:5",
         ]);
 
-        if($validator->fails()) {
+        if ($validator->fails()) {
             return redirect()->back()->withErrors($validator)->withInput()->with([
                 "localizacao_erro" => "salvar_campo",
                 "titulo_old" => $request->titulo,
                 "legenda_old" => $request->legenda,
             ]);
         }
-        
+
         $campo = new Campo;
-        if($request->campo_id != NULL) $campo = Campo::find($request->campo_id); //EDIÇÂO
+        if ($request->campo_id != NULL) $campo = Campo::find($request->campo_id); //EDIÇÂO
         $campo->fill($request->all());
         $campo->save();
         return redirect()->back();
     }
 
 
-    public function deletarCampo(Request $request) {
+    public function deletarCampo(Request $request)
+    {
         try {
             Campo::find($request->campo_id)->delete();
+        } catch (Exception $ex) {
         }
-        catch(Exception $ex) {}
         return redirect()->back();
     }
 
 
-    public function anotacoes_html($id_campo) {
+    public function anotacoes_html($id_campo)
+    {
         $campo = Campo::find($id_campo);
-        if($campo) return view('AtividadeAcademica.secao.listagem_anotacoes', ["campo" => $campo]);
+        if ($campo) return view('AtividadeAcademica.secao.listagem_anotacoes', ["campo" => $campo]);
     }
 
 
-    public function salvar_anotacao(Request $request) {
+    public function salvar_anotacao(Request $request)
+    {
         $anotacao = new Anotacao;
         $anotacao->fill($request->all());
         $anotacao->user_id = Auth::id();
         $anotacao->status = 1;
         $anotacao->save();
-        
+
         return "";
     }
 
-    public function deletar_anotacao(Request $request) {
-        try{
-            //TODO: permissoes sobre a anotação
+    public function deletar_anotacao(Request $request)
+    {
+        try {
             $anotacao = Anotacao::find($request->anotacao_id);
-            $anotacao->delete();
+            if ($anotacao->campo->secao->atividade->user_logado_proprietario() || $anotacao->user_id == Auth::id()) {
+                $anotacao->delete();
+            }
+        } catch (Exception $ex) {
         }
-        catch(Exception $ex) {}
     }
 
 
-    public function salvar_conteudo(Request $request) {
-            $campo = Campo::find($request->id_campo);
-            $campo->conteudo = $request->conteudo;
-            $campo->save();
-            return redirect()->back();
+    public function salvar_conteudo(Request $request)
+    {
+        $campo = Campo::find($request->id_campo);
+        $campo->conteudo = $request->conteudo;
+        $campo->save();
+        return redirect()->back();
     }
 }
